Fix issue with attention layer
Add separte training and model files for attention model
